test: cover block property transform
