+++ codemods/__testfixtures__/block-to-width-100.output.tsx
@@ -1,0 +1,8 @@
+import { Button } from '@freenow/wave';
+import React from 'react';
+
+export const ButtonTest = (): JSX.Element => (
+    <Button width="100%" mr="1" mt="3" disabled>
+        Clone
+    </Button>
+);
