+++ codemods/__tests__/block-to-width-100-test.ts
@@ -1,0 +1,4 @@
+jest.autoMockOff();
+const { defineTest } = require('jscodeshift/dist/testUtils');
+
+defineTest(__dirname, 'block-to-width-100', null, 'block-to-width-100', { parser: 'tsx' });
